Upgraded dependencies. Switched to TypeScript
