--- src/components/Notification.js
@@ -1,11 +1,0 @@
-import React from "react";
-import { Alert } from "react-bootstrap";
-const Notification = ({ text, type }) => {
-  return (
-    <Alert bsStyle={type}>
-      <p>{text}</p>
-    </Alert>
-  );
-};
-
-export default Notification;
